requirements.txt
numpy==1.24.3
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
Flask==3.0.2
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

openai==1.60.1
anthropic==0.8.1
python-dotenv==1.0.0
pillow==10.2.0  
tenacity==8.2.3 

python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6

base64io==1.0.3

selenium==4.27.1

scikit-learn==1.2.2
tensorflow
statsmodels